+++ src/main/java/net/mvndicraft/roadspeedmounts/MountsMoveListener.java
@@ -2,6 +2,7 @@ package net.mvndicraft.roadspeedmounts;
 
 import org.bukkit.attribute.Attributable;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerMoveEvent;
@@ -9,19 +10,19 @@ import org.bukkit.event.player.PlayerMoveEvent;
 public class MountsMoveListener implements Listener {
     @EventHandler(ignoreCancelled = true)
     public void onMountsMove(PlayerMoveEvent event) {
-        Entity mounts = event.getPlayer().getVehicle();
+        Player player = event.getPlayer();
+        Entity mounts = player.getVehicle();
         if (mounts != null) {
-            if (RoadSpeed.isPassengersAffected(mounts.getPassengers())) {
-                RoadSpeedMountsPlugin.debug(() -> "Passenger affected: " + mounts.getPassengers());
-                if (mounts instanceof Attributable attributable) {
-                    RoadSpeed.applyRoadSpeed(attributable, RoadSpeed.getRoadSpeed(mounts));
-                }
-            } else {
-                // Reset to 0 in case player has switch gamemode
-                if (mounts instanceof Attributable attributable) {
-                    RoadSpeed.applyRoadSpeed(attributable, 0);
-                }
-            }
+            applyRoadSpeed(mounts, !RoadSpeed.isPassengersAffected(mounts.getPassengers()));
+        } else {
+            applyRoadSpeed(player, !RoadSpeed.isAffected(player));
+        }
+    }
+
+    private void applyRoadSpeed(Entity entity, boolean toZero) {
+        if (entity instanceof Attributable attributable) {
+            RoadSpeedMountsPlugin.debug(() -> "Attributable affected: " + attributable);
+            RoadSpeed.applyRoadSpeed(attributable, toZero ? 0 : RoadSpeed.getRoadSpeed(entity));
         }
     }
 }